+++ src/application/use-cases/organization/delete.use-case.ts
@@ -12,6 +12,10 @@ export class DeleteOrganizationUseCase {
      * @returns A promise that resolves to a boolean indicating whether the organization was deleted successfully.
      */
     async execute(id: string): Promise<boolean> {
+        const organization = await this.organizationRepository.getById(id);
+        if (!organization) {
+            throw new Error(`Organization with ID ${id} not found`);
+        }
         return await this.organizationRepository.delete(id);
     }
 }
