+++ src/application/use-cases/schedule/delete.use-case.ts
@@ -5,6 +5,10 @@ export class DeleteUseCase {
     constructor(private scheduleRepository: ScheduleRepository) { }
 
     async execute(id: string): Promise<boolean> {
+        const schedule = await this.scheduleRepository.getById(id);
+        if (!schedule) {
+            throw new Error(`Schedule with ID ${id} not found`);
+        }
         return await this.scheduleRepository.delete(id);
     }
 }
